scale by mean of group mean for imbalanced data

scale by mean of group mean for imbalanced data

scale_mgm(expr, label)

Arguments

expr

matrix

label

a vector of group label

Value

scaled matrix

Details

$$z=\frac{x-\frac{\sum_k^{n_D}(\mu_k)}{n_D}}{s_{pooled}}$$ where \(s_{pooled}=\sqrt{\frac{\sum_k^{n_D}{(n_k-1){s_k}^2}}{\sum_k^{n_D}{n_k}-k}}\)

Examples

scale_mgm(matrix(rnorm(100), 10), label = rep(letters[1:2], 5))
#>             [,1]        [,2]        [,3]        [,4]       [,5]        [,6]
#>  [1,]  1.3087901 -0.46870958 -1.77434322 -0.73647719  0.3371481  0.32581901
#>  [2,] -1.3753000  0.10250723 -1.21830292 -0.10062851  0.1549302 -0.05315493
#>  [3,]  1.7617188 -0.24191322 -0.70636940 -0.98174123  1.3961940  0.89530742
#>  [4,] -0.3369481 -0.57365427 -1.16897077  1.64935095 -0.5560953  1.07063296
#>  [5,] -0.8908733  1.63819192  0.70859671  0.01519947 -2.0324866 -0.08205510
#>  [6,] -0.6104932 -1.06680499  0.86949708  0.36850181  1.5516198 -0.80070745
#>  [7,] -0.2463602  0.26426515 -2.25922177  1.51980452  0.6513791  1.14501482
#>  [8,] -1.5955237  0.07530608  0.34596634 -1.26371424 -0.5118640  0.02510103
#>  [9,] -1.2847095  0.74186234 -0.07197136  0.40590393  0.6022167  0.63325326
#> [10,] -0.4409084 -0.86314621 -0.54935865  0.81761713 -0.8240249  0.49646428
#>              [,7]        [,8]       [,9]      [,10]
#>  [1,]  0.15985425  1.21408209 -0.8342649  0.4681014
#>  [2,]  0.82067860 -0.82619451  0.8238921  1.6715727
#>  [3,] -0.70119006 -0.59587251 -0.5331031 -0.2930307
#>  [4,] -0.48834043  0.67411005  0.9420617 -1.2121468
#>  [5,]  0.08234043  0.14292776 -1.2772367  1.6953955
#>  [6,] -1.38479712 -0.42851144  0.9576904  0.5440051
#>  [7,] -0.71714376 -0.41107979 -1.5015499  1.5548919
#>  [8,]  0.90256015  0.09309651  1.6940613  0.2350105
#>  [9,] -0.46308406 -1.04828826  1.5618972 -1.0770802
#> [10,] -0.11260505 -1.23151969  0.9205192  1.7869622